fixed quiz
