+++ quiz.rb
@@ -1,0 +1,7 @@
+class Quiz
+  attr_accessor :prompt, :answer
+  def initialize(prompt, answer)
+    @prompt = prompt
+    @answer = answer
+  end
+end
